pyproject.toml
[project]
name = "vqa-colab-local-tools"
version = "0.1.0"
description = "Local utilities for VQA data inspection and notebook execution"
readme = "AGENTS.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "jupytext>=1.17.3",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
]

[tool.uv]
package = false
